src/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
google-auth==2.43.0
google-genai==1.53.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.3.5
ollama==0.6.1
openai==2.9.0
pandas==2.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
